added edited_at to messages + first socket implementation (to fix)

+++ src/server.ts
@@ -1,9 +1,9 @@
 import { createServer } from "http";
 import app from "./app";
 import jwt from "jsonwebtoken";
-import { ExtendedError, Server } from "socket.io";
-import { PrismaClient, Chat } from "../client";
-import { getEnvOrThrow } from "./controllers/helpers";
+import { Server } from "socket.io";
+import { PrismaClient, Chat, Message } from "../client";
+import { checkIfUserHasMessage, getEnvOrThrow } from "./controllers/helpers";
 
 const prisma = new PrismaClient();
 
@@ -16,39 +16,87 @@ const io = new Server(server, {
   },
 });
 
-const isSocketError = (obj: unknown): obj is ExtendedError => {
-  if ((obj as ExtendedError).data) {
-    return true;
+interface HandshakeAuthObject {
+  user: string;
+}
+
+const isAuthObject = (obj: unknown): obj is HandshakeAuthObject => {
+  if (obj) {
+    return typeof (obj as HandshakeAuthObject).user === "string";
   }
   return false;
 };
 
 io.use((socket, next) => {
-  const authToken = socket.handshake.query.token;
-
-  if (typeof authToken === "string") {
-    const decodedUser = jwt.verify(authToken, getEnvOrThrow("JWT_SECRET_KEY"));
-    if (typeof decodedUser === "object") {
-      socket.handshake.auth = { user: decodedUser.user_uuid };
-      return next();
+  try {
+    const authToken = socket.handshake.query.token;
+    console.log(authToken);
+    if (typeof authToken === "string") {
+      const decodedUser = jwt.verify(
+        authToken,
+        getEnvOrThrow("JWT_SECRET_KEY")
+      );
+      console.log(decodedUser);
+      if (typeof decodedUser === "object") {
+        console.log("updating handshake");
+        socket.handshake.auth = { user: decodedUser.user_uuid };
+        //console.log(socket.handshake.auth);
+        return next();
+      }
+      return next(new Error("Unauthorized"));
     }
     return next(new Error("Unauthorized"));
+  } catch (err) {
+    return next(new Error("Unauthorized"));
   }
-  return next(new Error("Unauthorized"));
 });
 
 io.on("connection", (socket) => {
-  console.log(socket);
+  //console.log(socket.id);
   socket.on("join", ({ room }) => {
+    //console.log(room);
     socket.join(room);
   });
   socket.on(
     "delete chat",
     async ({ room, chats }: { room: string; chats: Chat[] }) => {
-      //console.log(socket.handshake);
       try {
-        //console.log(room);
-        //console.log("delete chat");
+        if (!isAuthObject(socket.handshake.auth)) {
+          throw new Error("Unauthorized");
+        }
+        const authUser = socket.handshake.auth.user;
+
+        const authUserChats = await prisma.user.findUniqueOrThrow({
+          where: {
+            uuid: authUser,
+          },
+          select: {
+            chats: {
+              where: {
+                Chat: {
+                  uuid: room,
+                },
+              },
+              select: {
+                Chat: {
+                  select: {
+                    uuid: true,
+                  },
+                },
+              },
+            },
+          },
+        });
+
+        const chatFound = authUserChats.chats.find(
+          (chat) => chat.Chat.uuid === room
+        );
+        //console.log(chatFound, authUserChats, authUser);
+
+        if (!chatFound) {
+          throw new Error("chat not found");
+        }
+
         const deletedChat = await prisma.chat.delete({
           where: {
             uuid: room,
@@ -66,8 +114,207 @@ io.on("connection", (socket) => {
         );
 
         io.to(room).emit("chat deleted", { updatedChats });
+      } catch (err: unknown) {
+        //console.log("chat error");
+        socket.emit("chat error", {
+          error:
+            err instanceof Error
+              ? err.message
+              : "error while deleting the chat",
+        });
+      }
+    }
+  );
+  socket.on("send message", async ({ chatUUID, receiverUUID, content }) => {
+    try {
+      if (!isAuthObject(socket.handshake.auth)) {
+        throw new Error("Unauthorized");
+      }
+      const senderUUID = socket.handshake.auth.user;
+      //the a socket room identifier is the same as the chat identifier
+
+      const sender = await prisma.user.findUniqueOrThrow({
+        where: {
+          uuid: senderUUID,
+        },
+      });
+
+      const receiver = await prisma.user.findUniqueOrThrow({
+        where: {
+          uuid: receiverUUID,
+        },
+      });
+
+      const userChats = await prisma.userChat.findMany({
+        where: {
+          user: {
+            uuid: {
+              in: [senderUUID, receiverUUID],
+            },
+          },
+        },
+        select: {
+          Chat: {
+            select: {
+              id: true,
+              uuid: true,
+            },
+          },
+        },
+      });
+
+      const chatFound = userChats.find((chat) => chat.Chat.uuid === chatUUID);
+
+      if (!chatFound) {
+        throw new Error("chat not found");
+      }
+
+      if (chatFound) {
+        const newMessageData = {
+          sender_id: sender?.id,
+          receiver_id: receiver?.id,
+          chat_id: chatFound.Chat.id,
+          content,
+        };
+        const newMessage = await prisma.message.create({
+          data: newMessageData,
+        });
+
+        const { uuid, content: newMessageContent, created_at } = newMessage;
+        io.to(chatUUID).emit("message created", {
+          message: {
+            uuid,
+            content: newMessageContent,
+            created_at,
+            senderUUID: senderUUID,
+          },
+        });
+      }
+    } catch (err) {
+      socket.emit("message error", {
+        error:
+          err instanceof Error ? err.message : "error while deleting the chat",
+      });
+    }
+
+    //console.log(socket.handshake.auth, chatUUID, receiver, content);
+  });
+  socket.on(
+    "delete message",
+    async ({
+      room,
+      message,
+      status,
+    }: {
+      room: string;
+      message: string;
+      status: string;
+    }) => {
+      //console.log("delete message");
+      try {
+        if (!isAuthObject(socket.handshake.auth)) {
+          throw new Error("Unauthorized");
+        }
+
+        const authUser = socket.handshake.auth.user;
+
+        const messageFound = await checkIfUserHasMessage(
+          authUser,
+          message,
+          status
+        );
+        if (!messageFound) {
+          throw new Error("message not found");
+        }
+
+        const deletedMessage = await prisma.message.delete({
+          where: {
+            uuid: message,
+          },
+          select: {
+            uuid: true,
+          },
+        });
+
+        io.to(room).emit("message deleted", {
+          messageUUID: deletedMessage.uuid,
+        });
+      } catch (err: unknown) {
+        //console.log("chat error");
+        socket.emit("message error", {
+          error:
+            err instanceof Error
+              ? err.message
+              : "error while deleting the chat",
+        });
+      }
+    }
+  );
+  socket.on(
+    "edit message",
+    async ({
+      room,
+      message,
+      newMessage,
+      status,
+    }: {
+      room: string;
+      message: string;
+      newMessage: string;
+      status: string;
+    }) => {
+      try {
+        if (!isAuthObject(socket.handshake.auth)) {
+          throw new Error("Unauthorized");
+        }
+        const authUser = socket.handshake.auth.user;
+        console.log(room, message, newMessage, status, authUser);
+
+        const messageFound = await checkIfUserHasMessage(
+          authUser,
+          message,
+          status
+        );
+
+        if (!messageFound) {
+          throw new Error("message not found");
+        }
+
+        const updatedMessage = await prisma.message.update({
+          where: {
+            uuid: message,
+          },
+          data: {
+            content: newMessage,
+            edited_at: new Date(Date.now()),
+          },
+          select: {
+            uuid: true,
+            content: true,
+            created_at: true,
+            sender: {
+              select: {
+                uuid: true,
+              },
+            },
+          },
+        });
+
+        const { sender, ...rest } = updatedMessage;
+
+        io.to(room).emit("message updated", {
+          message,
+          updatedMessage: { ...rest },
+          senderUUID: sender?.uuid,
+        });
       } catch (err) {
-        io.to(room).emit("couldn't delete the chat");
+        console.log(err);
+        socket.emit("message error", {
+          error:
+            err instanceof Error
+              ? err.message
+              : "error while deleting the chat",
+        });
       }
     }
   );
